Feat : Notice service for admin
생성, 수정, 삭제 권한 확인 기능 추가

+++ server/src/main/java/toyspringboot/server/Service/NoticeService.java
@@ -12,6 +12,7 @@ import toyspringboot.server.Domain.Entity.QnA;
 import toyspringboot.server.Domain.Repository.NoticeRepository;
 import toyspringboot.server.Domain.Repository.UserRepository;
 
+import static org.springframework.http.HttpStatus.FORBIDDEN;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 
 @Service
@@ -23,19 +24,21 @@ public class NoticeService {
 
     public NoticeDto createNotice(String userToken, NoticeDto noticeDto) {
         UserDto userDto = UserDto.of(userRepository.findByEmail(userToken).orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "존재하지 않는 사용자 입니다.")));
+        if(!userDto.getAdmin()) throw new ResponseStatusException(FORBIDDEN, "공지 생성 권한이 없습니다.");
         noticeDto.setCreateNoticeDto(userDto);
         Notice notice = NoticeDto.toEntity(noticeDto);
         return NoticeDto.of(noticeRepository.save(notice));
     }
 
     public NoticeDto readNotice(String userToken, NoticeDto noticeDto) {
-        UserDto userDto = UserDto.of(userRepository.findByEmail(userToken).orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "존재하지 않는 사용자 입니다.")));
+        UserDto userDto = UserDto.of(userRepository.findByEmail(userToken).orElseThrow(() -> new ResponseStatusException(FORBIDDEN, "존재하지 않는 사용자 입니다.")));
         Notice notice = noticeRepository.findById(noticeDto.getId()).orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "존재하지 않는 공지 입니다."));
         return NoticeDto.of(notice);
     }
 
     public NoticeDto updateNotice(String userToken, NoticeDto noticeDto) {
         UserDto userDto = UserDto.of(userRepository.findByEmail(userToken).orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "존재하지 않는 사용자 입니다.")));
+        if(!userDto.getAdmin()) throw new ResponseStatusException(FORBIDDEN, "공지 수정 권한이 없습니다.");
         Notice notice = noticeRepository.findById(noticeDto.getId()).orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "존재하지 않는 공지 입니다."));;
         noticeRepository.updateNotice(notice, noticeDto);
         return NoticeDto.of(notice);
@@ -43,6 +46,7 @@ public class NoticeService {
 
     public NoticeDto deleteNotice(String userToken, NoticeDto noticeDto) {
         UserDto userDto = UserDto.of(userRepository.findByEmail(userToken).orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "존재하지 않는 사용자 입니다.")));
+        if(!userDto.getAdmin()) throw new ResponseStatusException(FORBIDDEN, "공지 삭제 권한이 없습니다.");
         Notice notice = noticeRepository.findById(noticeDto.getId()).orElseThrow(() -> new ResponseStatusException(NOT_FOUND, "존재하지 않는 공지 입니다."));;
         noticeRepository.deleteNotice(notice, noticeDto);
         return NoticeDto.of(notice);
